--- pip install
@@ -1,1 +1,0 @@
-pandas
