+++ src/pip install
@@ -1,0 +1,1 @@
+pandas
